api/pyproject.toml
[tool.poetry]
name = "mesh"
version = "0.1.0"
description = "Python Mesh App"
authors = ["Eduardo Garcia <24eg14@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.1"
uvicorn = "^0.23.1"
pydantic = "^2.1.1"
requests = "^2.31.0"
sqlalchemy = "^2.0.19"
python-dotenv = "^1.0.0"
alembic = "^1.11.1"
pg8000 = "^1.30.1"
starlette = "0.36.3"
strawberry-graphql = "^0.224.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
sqlalchemy2-stubs = "^0.0.2a35"
requests = "^2.31.0"
mypy = "^1.7.1"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
